+++ CHANGELOG.md
@@ -19,7 +19,10 @@ Copyright (c) 2022-2026 Claudio Satriano <satriano@ipgp.fr>
 - New command `seiscat run` to run a user-defined command on each event
 - New command `seiscat samplescript` to write a sample script for the
   `seiscat run` command
-- New GeoJSON output format for `seiscat print` command (`-f geojson`)
+- New command `seiscat export` to export the catalog to a file in
+  a specified format (currently supported formats are CSV and GeoJSON)
+- `seiscat print`: remove the CSV format, which is now available
+  through `seiscat export`
 - New commands `seiscat get` and `seiscat set` to get and set the value of
   a specfic event attribute
 - New plottype `plotly` for `seiscat plot`, producing interactive 3D seismicity